prefetch-npm-deps: move tests to separate file
